+++ proyectobanco/src/App.js
@@ -1,18 +1,23 @@
 import './App.css';
-import { Header } from './components/Header';
+import { Header } from './components/header/Header';
 import 'bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { Home } from './components/Home';
+import { Footer } from './components/footer/Footer';
 function App() {
   return (
     <>
+      <Header />
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<Home />}></Route>
           <Route path="/listado" element={<Header />}></Route>
         </Routes>
       </BrowserRouter>
+      <div>
+        <Footer />
+      </div>
     </>
   );
 }
