--- proyectobanco/src/components/estilos.css
@@ -1,3 +1,0 @@
-.navbarM {
-  background-color: #1b2d54;
-}
