+++ proyectobanco/src/components/footer/Footer.jsx
@@ -1,0 +1,16 @@
+import './footer.css'
+export function Footer() {
+    return (
+        <>
+            <footer className="bg-light text-center text-lg-start">
+
+                <div className="text-center p-3  foot" >
+                    <div>
+                        <h5 className="colorText ms-5"> Dirección General: Avenida de las Américas # 42-81 PBX: (601) 650 1050 </h5>
+                    </div>
+                </div>
+
+            </footer>
+        </>
+    );
+}
