+++ proyectobanco/src/components/footer/footer.css
@@ -1,0 +1,16 @@
+.foot {
+  display: flex;
+  background-color: #1b2d54;
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  height: 40px;
+}
+.colorText {
+  display: flex;
+  color: whitesmoke !important;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center !important;
+  margin-left: 30rem !important;
+}
